refactor(extension): 설정 스토어 분리 및 widget 레이어로 이전
