+++ apps/extension/src/app/entrypoint/side-panel/SidePanel.tsx
@@ -1,21 +1,19 @@
 import { SIDE_PANEL_SCREEN_MAP } from "@/app/entrypoint/side-panel/Navigation";
-import { useAuthStore, useSettingStore } from "@/app/store/model";
+import { useAuthStore } from "@/entities/auth/model";
 import AuthScreen from "@/screens/auth/ui/AuthScreen";
-import {
-  SidePanelFooter,
-  SidePanelLayout,
-  SidePanelTabs,
-} from "@/widgets/layout/ui";
+import { SidePanelFooter, SidePanelLayout } from "@/widgets/layout/ui";
+import { useTabNavigationStore } from "@/widgets/tab-navigation/model";
+import { TabNavigation } from "@/widgets/tab-navigation/ui";
 
 export function SidePanel() {
-  const activeTab = useSettingStore((state) => state.activeTab);
+  const activeTab = useTabNavigationStore((state) => state.activeTab);
   const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
   const MainScreen =
     SIDE_PANEL_SCREEN_MAP[activeTab as keyof typeof SIDE_PANEL_SCREEN_MAP];
 
   return (
     <div className="flex h-full flex-col">
-      <SidePanelTabs />
+      <TabNavigation />
       <SidePanelLayout>
         {isLoggedIn ? <MainScreen /> : <AuthScreen />}
       </SidePanelLayout>
