--- apps/extension/src/app/store/model/index.ts
@@ -1,2 +1,0 @@
-export * from "./auth.store";
-export * from "./setting.store";
