+++ apps/extension/src/entities/auth/model/index.ts
@@ -1,0 +1,1 @@
+export * from "./auth.store";
