+++ apps/extension/src/screens/ai-recap/ui/AiRecapScreen.tsx
@@ -1,17 +1,10 @@
-import { useSettingStore } from "@/app/store/model";
 import { AiRecapContent } from "@/features/ai-recap/ui";
-import { DatePicker } from "@/shared/ui";
+import DateSelector from "@/widgets/date-selector/ui/DateSelector";
 
 const AiRecapScreen = () => {
-  const selectedDate = useSettingStore((state) => state.selectedDate);
-  const setSelectedDate = useSettingStore((state) => state.setSelectedDate);
-
   return (
     <>
-      <DatePicker
-        value={selectedDate ?? new Date()}
-        onChange={(date) => setSelectedDate(date ?? new Date())}
-      />
+      <DateSelector />
 
       <AiRecapContent />
     </>
