+++ apps/extension/src/widgets/date-selector/model/index.ts
@@ -1,0 +1,1 @@
+export * from "./date-selector.store";
