+++ apps/extension/src/widgets/layout/ui/SidePanelLayout.tsx
@@ -6,9 +6,9 @@ const SidePanelLayout = ({
   className,
 }: PropsWithChildren<{ className?: string }>) => {
   return (
-    <div className={cn("pt-12 pb-[7rem] w-full flex-1 ", className)}>
+    <main className={cn("pt-12 pb-[7rem] w-full flex-1 ", className)}>
       {children}
-    </div>
+    </main>
   );
 };
