+++ apps/extension/src/widgets/layout/ui/index.ts
@@ -1,3 +1,2 @@
 export { default as SidePanelFooter } from "./SidePanelFooter";
 export { default as SidePanelLayout } from "./SidePanelLayout";
-export { default as SidePanelTabs } from "./SidePanelTabs";
