+++ apps/extension/src/widgets/tab-navigation/model/index.ts
@@ -1,0 +1,1 @@
+export * from "./tab-navigation.store";
