+++ apps/extension/src/widgets/tab-navigation/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as TabNavigation } from "./TabNavigation";
